8: Rename examples to demos.
