+++ src/components/demos/geoms/SpecialPolygons.vue
