+++ src/components/demos/geoms/index.js
@@ -1,0 +1,11 @@
+import Geometries from '@/components/demos/geoms/Geometries'
+import SpecialPolygons from '@/components/demos/geoms/SpecialPolygons'
+import GeocentricLine from '@/components/demos/geoms/GeocentricLine'
+
+let demos = [  
+    Geometries,
+    SpecialPolygons,
+    GeocentricLine
+]
+
+export default demos;
