+++ src/components/demos/index.js
@@ -1,0 +1,7 @@
+import HelloMap from '@/components/demos/HelloMap'
+import LoadShapefile from '@/components/demos/LoadShapefile'
+import GeomDemos from './geoms';
+
+export default [
+    HelloMap, LoadShapefile, ...GeomDemos
+];
