--- src/components/examples/geoms/index.js
@@ -1,11 +1,0 @@
-import Geometries from '@/components/examples/geoms/Geometries'
-import SpecialPolygons from '@/components/examples/geoms/SpecialPolygons'
-import GeocentricLine from '@/components/examples/geoms/GeocentricLine'
-
-let demos = [  
-    Geometries,
-    SpecialPolygons,
-    GeocentricLine
-]
-
-export default demos;
