--- src/components/examples/index.js
@@ -1,7 +1,0 @@
-import HelloMap from '@/components/examples/HelloMap'
-import LoadShapefile from '@/components/examples/LoadShapefile'
-import GeomDemos from './geoms';
-
-export default [
-    HelloMap, LoadShapefile, ...GeomDemos
-];
